utils/requirements.txt
requests
blurhash-python
